Text rendering can now have the color controlled per character. Eventually this will support color Lerping as well.
Moved more globals into GameState, and made Input and Console into file local storage instead of singleton.

+++ GLMario/source/utility.h
@@ -66,8 +66,8 @@ T* ExpandArray(T* array, uint32 start_count, uint32 new_count)
 
 struct MemoryArena
 {
-    uint32 size;
-    uint32 used;
+    size_t size;
+    size_t used;
     uint8* base;
 };
